Include newest build. (remove redundant cast causing error)

+++ code/src/asl/sensor/DataBlock.java
@@ -189,7 +189,7 @@ public class DataBlock {
   }
 
   public long getEndTime() {
-    return startTime + ( (long) interval * data.size() );
+    return startTime + ( interval * data.size() );
   }
   
 }
